bugfix. fix the excercises not copied properly

+++ python_app/outside_calls/fsApp_calls.py
@@ -1,8 +1,8 @@
-import subprocess
 import os
 import shutil
+import pathlib
+from distutils.dir_util import copy_tree
 
-import time
 import _utils._utils_main as _u
 import _utils.logging as log
 import settings.facade as sf
@@ -20,7 +20,17 @@ class FinderCalls:
         cls.copyFile(sourceFilepath, destFilepath)
         cls.deleteFile(sourceFilepath)
 
-    def copyFile(sourceFilepath, destFilepath):
+    @classmethod
+    def copyFile(cls, sourceFilepath, destFilepath):
+        sourceFilepath_pl = pathlib.Path(sourceFilepath)
+        destFolder = pathlib.Path(destFilepath).parent
+
+        if not cls.checkIfFileOrDirExists(destFolder):
+            os.makedirs(destFolder, exist_ok=True)
+
+        if not sourceFilepath_pl.is_file():
+            return copy_tree(sourceFilepath, destFilepath)
+
         return shutil.copy2(sourceFilepath, destFilepath)
 
     def moveFolder(sourceFilepath, destFilepath):
